normalized return sentence analyzer AP

+++ app/src/access_points/analyze/sentence.py
@@ -7,9 +7,9 @@ def analyzer(sentence, sentenceAnalyzer):
     try:
         analysis = sentenceAnalyzer.analyze(sentence)
         results = analysis['res']
-        data = {'positivity': round(results[2], 4),
-                'neutrality': round(results[1], 4),
-                'negativity': round(results[0], 4)}
+        data = {'positivity': float(round(results[2], 4)),
+                'neutrality': float(round(results[1], 4)),
+                'negativity': float(round(results[0], 4))}
         g.user.save_query('/analyze/sentence', data, analysis['exec_time'])
         return jsonify(data)
     except Exception as e:
